admin auth + calc page + admin AJAX
Fixing up the admin/login and admin/register pages so they're the same visually as the main login and register pages. Updated the main calculation page so that it has fields for job name and job description, and that the method names are aligned with the checkboxes. Implemented AJAX call on admin analytics page to pull live counts of calculations in progress.