database update - seeders + redshift.user_id, request update
Updated seeders to reflect the feedback from client about expected redshift value ranges from 0-10. Also updated the database schema to include a field that marks whether a method has been deleted (for soft deletion), updated the redshifts table to exclude user_id (as that is contained within the jobs table that it already links to via job_id), and added redshift_alt_result in the calculations table so that non-numeric (e.g. file locations) can be written to the database. Also changed the format of single galaxy requests so that they match the format of multiple galaxy requests, where method_id, user_id, and the token are sent as the last line/array inside the JSON