Added complete and per-job exports of results, including files
Re-introduced result export to .csv after it was removed at some point. All files download as a .zip, and if there are any redshift_alt_results associated with a galaxy (these should be files), then those files will be included in the zip file and their filename included in the master results .csv file also in the .zip. After download, files are deleted from the server to avoid clutter