Updated analytics page, methods uploading, more api checks, guest page, history page, system load widget
Updated analytics page so it pulls the correct labels for the correct data. Updated method uploading logic based on failed test cases to include checking a method name is entered when creating/updating, and that the method file is .py. Added some additional checks to the API connection, so it checks for connection status first, then checks if it receives a response of anything but 200. Guest page is revamped, now stores no data on database and request is sent from on the guestresult page in AJAX with the help of session variables. Also made some small changes to the history page to make links look like links, and added the first workings of the system load widget on the admin panel dashboard